+++ content/Bat/Bat.Api/Services/IAuthenticator.cs
@@ -1,19 +1,26 @@
-﻿using Bat.Api.Controllers;
-using Bat.Shared.Api;
+﻿using Bat.Shared.Api;
+using Bat.Shared.ExternalLoginHelper;
 
 namespace Bat.Api.Services;
 
 public interface IAuthenticator
 {
 	/// <summary>
-	/// Perform an authentication action.
+	/// Performs an authentication action.
 	/// </summary>
 	/// <param name="req">The authentication request.</param>
 	/// <returns></returns>
 	public AuthResp Authenticate(AuthReq req);
 
 	/// <summary>
-	/// Refresh an issued authentication token.
+	/// Logs in a user.
+	/// </summary>
+	/// <param name="extProfile">The external user profile.</param>
+	/// <returns></returns>
+	public AuthResp Authenticate(ExternalUserProfile extProfile);
+
+	/// <summary>
+	/// Refreshes an issued authentication token.
 	/// </summary>
 	/// <param name="token">The issued authentication token.</param>
 	/// <param name="ignoreTokenSecurityCheck">If true, donot check if token's security tag is still valid.</param>
@@ -21,7 +28,7 @@ public interface IAuthenticator
 	public AuthResp Refresh(string token, bool ignoreTokenSecurityCheck = false);
 
 	/// <summary>
-	/// Validate an issued authentication token.
+	/// Validates an issued authentication token.
 	/// </summary>
 	/// <param name="token"></param>
 	/// <returns></returns>
@@ -31,14 +38,21 @@ public interface IAuthenticator
 public interface IAuthenticatorAsync
 {
 	/// <summary>
-	/// Perform an authentication action.
+	/// Performs an authentication action.
 	/// </summary>
 	/// <param name="req">The authentication request.</param>
 	/// <returns></returns>
 	public Task<AuthResp> AuthenticateAsync(AuthReq req);
 
 	/// <summary>
-	/// Refresh an issued authentication token.
+	/// Logs in a user.
+	/// </summary>
+	/// <param name="extProfile">The external user profile.</param>
+	/// <returns></returns>
+	public Task<AuthResp> AuthenticateAsync(ExternalUserProfile extProfile);
+
+	/// <summary>
+	/// Refreshes an issued authentication token.
 	/// </summary>
 	/// <param name="token">The issued authentication token.</param>
 	/// <param name="ignoreTokenSecurityCheck">If true, donot check if token's security tag is still valid.</param>
@@ -46,7 +60,7 @@ public interface IAuthenticatorAsync
 	public Task<AuthResp> RefreshAsync(string token, bool ignoreTokenSecurityCheck = false);
 
 	/// <summary>
-	/// Validate an issued authentication token.
+	/// Validates an issued authentication token.
 	/// </summary>
 	/// <param name="token"></param>
 	/// <returns></returns>