+++ content/Bat/Bat.Blazor/Bat.Blazor.App/Bootstrap/P1000AliveSyncBootstrapper.cs
@@ -1,6 +1,8 @@
 ﻿using Bat.Blazor.App.Helpers;
+using Bat.Blazor.App.Services;
 using Bat.Shared.Api;
 using Bat.Shared.Bootstrap;
+using Microsoft.AspNetCore.Components.Authorization;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using System.Text.Json;
@@ -110,7 +112,8 @@ public abstract class BackgroundTimerService : IDisposable
 public sealed class InfoSyncService : BackgroundTimerService
 {
 	private static readonly TimeSpan initialDelay = TimeSpan.Zero; // TimeSpan.Zero means start immediately
-	private static readonly TimeSpan interval = TimeSpan.FromMinutes(60);
+
+	private static readonly TimeSpan interval = TimeSpan.FromMinutes(30);
 
 	private readonly IServiceProvider serviceProvider;
 	private readonly ILogger<InfoSyncService> logger;
@@ -125,7 +128,7 @@ public sealed class InfoSyncService : BackgroundTimerService
 
 	protected override async void DoWork(object? state)
 	{
-		logger.LogDebug("{service} - pinging server...", GetType().FullName);
+		logger.LogInformation("{service} - pinging server...", GetType().FullName);
 		using (var scope = serviceProvider.CreateScope())
 		{
 			var apiClient = scope.ServiceProvider.GetRequiredService<IApiClient>();
@@ -140,6 +143,8 @@ public sealed class InfoSyncService : BackgroundTimerService
 				Globals.ServerInfo = infoResp.Data?.Server;
 				Globals.CryptoInfo = infoResp.Data?.Crypto;
 				Globals.Ready = true;
+				var stateContainer = scope.ServiceProvider.GetRequiredService<StateContainer>();
+				stateContainer.NotifyStateChanged();
 			}
 		}
 	}
@@ -152,8 +157,8 @@ public sealed class InfoSyncService : BackgroundTimerService
 /// </summary>
 public sealed class AuthTokenRefresherService : BackgroundTimerService
 {
-	private static readonly TimeSpan initialDelay = TimeSpan.Zero; // TimeSpan.Zero means start immediately
-	private static readonly TimeSpan interval = TimeSpan.FromSeconds(15);
+	private static readonly TimeSpan initialDelay = TimeSpan.FromMinutes(5);
+	private static readonly TimeSpan interval = TimeSpan.FromMinutes(30);
 
 	private readonly IServiceProvider serviceProvider;
 	private readonly ILogger<AuthTokenRefresherService> logger;
@@ -168,12 +173,34 @@ public sealed class AuthTokenRefresherService : BackgroundTimerService
 
 	protected override async void DoWork(object? state)
 	{
-		logger.LogDebug("{service} - refreshing auth token...", GetType().FullName);
+		logger.LogInformation("{service} - refreshing auth token...", GetType().FullName);
 		using (var scope = serviceProvider.CreateScope())
 		{
 			var localStorage = scope.ServiceProvider.GetRequiredService<LocalStorageHelper>();
 			var authToken = await localStorage.GetItemAsync<string>(Globals.LOCAL_STORAGE_KEY_AUTH_TOKEN);
-			logger.LogWarning("{service} - Auth Token: {token}", GetType().FullName, authToken);
+			if (string.IsNullOrEmpty(authToken)) return; // do nothing if auth token does not exist
+
+			if (string.IsNullOrEmpty(Globals.ApiBaseUrl))
+			{
+				logger.LogWarning("{service} - API base URL is not set.", GetType().FullName);
+				return;
+			}
+			var apiClient = scope.ServiceProvider.GetRequiredService<IApiClient>();
+			var authResp = await apiClient.RefreshAsync(authToken, Globals.ApiBaseUrl);
+			if (authResp.Status != 200)
+			{
+				logger.LogError("{service} - error refreshing auth token: {result}", GetType().FullName, authResp.Message);
+				if (authResp.Status < 500)
+				{
+					// auth token is invalid, remove it from local storage
+					await localStorage.RemoveItemAsync(Globals.LOCAL_STORAGE_KEY_AUTH_TOKEN);
+					var authenticationStateProvider = scope.ServiceProvider.GetRequiredService<AuthenticationStateProvider>();
+					((JwtAuthenticationStateProvider)authenticationStateProvider).NotifyStageChanged();
+				}
+				return;
+			}
+			logger.LogInformation("{service} - auth token refreshed.", GetType().FullName);
+			await localStorage.SetItemAsync(Globals.LOCAL_STORAGE_KEY_AUTH_TOKEN, authResp.Data.Token!);
 		}
 	}
 }