+++ content/Bat/Bat.Blazor/Bat.Blazor.App/Pages/Login.razor.cs
@@ -5,6 +5,7 @@ using Bat.Blazor.App.Shared;
 using Bat.Shared.Api;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Authorization;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -23,9 +24,12 @@ public partial class Login : BaseComponent
 	private string Email { get; set; } = string.Empty;
 	private string Password { get; set; } = string.Empty;
 
+	private IEnumerable<string> ExternalAuthProviders { get; set; } = [];
+
 	private string AlertType { get; set; } = "info";
 	private string AlertMessage { get; set; } = string.Empty;
 	private bool HideLoginForm { get; set; } = false;
+	private bool DisableExternalLogin { get; set; } = false;
 
 	[Inject]
 	private AuthenticationStateProvider AuthenticationStateProvider { get; set; } = default!;
@@ -46,25 +50,73 @@ public partial class Login : BaseComponent
 	[Inject]
 	private ILogger<Login> Logger { get; set; } = default!;
 
-	protected override async Task OnInitializedAsync()
+	protected override async Task OnAfterRenderAsync(bool firstRender)
 	{
-		ShowAlert("info", "Please wait...");
-		await base.OnInitializedAsync();
-		CloseAlert();
-
+		await base.OnAfterRenderAsync(firstRender);
 		// FIXME: NOT TO USE THIS IN PRODUCTION!
 		// for demo purpose: automatically fill the login form
-		if (Environment.GetEnvironmentVariable("DOTNET_RUNNING_IN_CONTAINER") == "true")
+		// if (Environment.GetEnvironmentVariable("DOTNET_RUNNING_IN_CONTAINER") == "true")
+		if (firstRender && HostEnvironment.Equals(EnvironmentName.Development, StringComparison.InvariantCultureIgnoreCase))
 		{
-			Logger.LogCritical("Running in container - automatically fill the login form. DO NOT USE THIS IN PRODUCTION!");
+			Logger.LogCritical("DevMode - automatically fill the login form. DO NOT USE THIS IN PRODUCTION!");
 			var seedUsers = await ApiClient.GetSeedUsersAsync(ApiBaseUrl);
 			Logger.LogCritical("Seed users: {users}", JsonSerializer.Serialize(seedUsers));
 			var user = seedUsers.Data?.FirstOrDefault();
 			Email = user?.Email ?? string.Empty;
 			Password = user?.Password ?? string.Empty;
 
-			ShowAlert("info", "DEMO ENV: The login form is automatically filled for your convenience.");
+			ShowAlert("info", "DevMode: Automatically fill login info.");
 		}
+	}
+
+	protected override async Task OnInitializedAsync()
+	{
+		ShowAlert("waiting", "Please wait...");
+		await base.OnInitializedAsync();
+
+		var providers = await ApiClient.GetExternalAuthProvidersAsync(ApiBaseUrl);
+		ExternalAuthProviders = providers.Data ?? [];
+
+		CloseAlert();
+	}
+
+	private async void BtnClickExternalLogin(string provider)
+	{
+		HideLoginForm = DisableExternalLogin = true;
+		ShowAlert("waiting", "Redirecting to external login provider...");
+		var returnUrl = QueryHelpers.ParseQuery(NavigationManager.ToAbsoluteUri(NavigationManager.Uri).Query)
+			.TryGetValue("returnUrl", out var returnUrlValue) ? returnUrlValue.FirstOrDefault("/") : "/";
+
+		ApiResp<string> apiResult;
+		switch (provider)
+		{
+			case "Microsoft":
+				var uriBuilder = new UriBuilder(NavigationManager.BaseUri)
+				{
+					Path = UIGlobals.ROUTE_LOGIN_EXTERNAL_MICROSOFT,
+					Query = QueryHelpers.AddQueryString(string.Empty, "returnUrl", returnUrl)
+				};
+				apiResult = await ApiClient.GetExternalAuthUrlAsync(new ExternalAuthUrlReq()
+				{
+					Provider = provider,
+					RedirectUrl = uriBuilder.ToString()
+				}, ApiBaseUrl);
+				break;
+			default:
+				HideLoginForm = DisableExternalLogin = false;
+				ShowModalNotImplemented();
+				return;
+		}
+
+		if (apiResult.Status != 200)
+		{
+			HideLoginForm = DisableExternalLogin = false;
+			ShowAlert("danger", $"{apiResult.Status}: {apiResult.Message ?? "Failed to get external auth URL."}");
+			return;
+		}
+
+		await Task.Delay(100); // UI hack to have the alert displayed before redirecting
+		NavigationManager.NavigateTo(apiResult.Data ?? UIGlobals.ROUTE_HOME);
 	}
 
 	private async void BtnClickLogin()
@@ -75,8 +127,8 @@ public partial class Login : BaseComponent
 			return;
 		}
 
-		HideLoginForm = true;
-		ShowAlert("info", "Authenticating, please wait...");
+		HideLoginForm = DisableExternalLogin = true;
+		ShowAlert("waiting", "Authenticating, please wait...");
 		var req = new AuthReq()
 		{
 			Email = Email,
@@ -85,12 +137,12 @@ public partial class Login : BaseComponent
 		var resp = await ApiClient.LoginAsync(req, ApiBaseUrl);
 		if (resp.Status != 200)
 		{
-			HideLoginForm = false;
+			HideLoginForm = DisableExternalLogin = false;
 			ShowAlert("danger", resp.Message!);
 			return;
 		}
 
-		ShowAlert("success", "Authenticated successfully, logging in...");
+		ShowAlert("success", "Authenticated successfully, logging in...⏳");
 		var returnUrl = QueryHelpers.ParseQuery(NavigationManager.ToAbsoluteUri(NavigationManager.Uri).Query)
 			.TryGetValue("returnUrl", out var returnUrlValue) ? returnUrlValue.FirstOrDefault("/") : "/";