+++ RELEASE-NOTES.md
@@ -1,5 +1,25 @@
 # Blazor Admin Template release notes
 
+## 2024-12-02 - v1.2.0
+
+### Added/Refactoring
+
+- Feature: Login with external account - Microsoft.
+- Refactor template structure - added new project `Libs`.
+
+### Fixed/Improvement
+
+- Fix(CodeQL): Nested if statements.
+- Fix(others): Minor bug fixes and improvements.
+
+## 2024-11-21 - v1.1.1
+
+### Fixed/Improvement
+
+- Fix(deps): Prevent dependencies from jumping to .NET 9.
+- Improvement: Notify stage changed when AppInfo is fetched.
+- Fix(dev/demo): Login/Profile info are automatically filled on Development env.
+
 ## 2024-10-31 - v1.1.0
 
 ### Added/Refactoring